backend_AI/requirements.txt
# 기본 웹 프레임워크
flask==3.0.0

# 환경 변수 관리
python-dotenv==1.0.0

# LangChain 관련
langchain==0.1.0
langchain-core==0.1.13
langchain-community==0.0.13
langchain-openai==0.0.5

# OpenAI
openai==1.12.0

# 벡터 저장소
faiss-cpu==1.7.4

# 유틸리티
numpy==1.24.3